Move getCaseIdParts to common
Jira ticket: CAMS-288
